Stop using instances of Client, and use class methods instead.
Httparty does weird things with class variables, and trying to
instantiate an instance of a class that mixes it in is a bad idea. Our
initialize method was setting class variables, which would lead to very
unexpected behavior if more than one was instantiated.

This also makes it easier for us to extract various API calls into a
canonical top-level method, without having to track a particular
instance of Client.